pyproject.toml
[project]
name = "neuralnoise"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Leonardo Piñeyro", email = "leopiney@gmail.com" }
]
requires-python = ">=3.10,<3.12"
dependencies = [
    "autogen-agentchat~=0.2",
    "elevenlabs>=1.10.0",
    "openai>=1.52.2",
    "pydantic>=2.9.2",
    "pydub>=0.25.1",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "typer>=0.12.5",
]

[project.scripts]
nn = "neuralnoise.cli:app"

[project.optional-dependencies]
local = [
    "docker>=7.1.0",
    "ollama>=0.3.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/neuralnoise"]

[tool.uv]
dev-dependencies = [
    "ipdb>=0.13.13",
    "ipykernel>=6.29.5",
    "ipython>=8.28.0",
    "mypy>=1.13.0",
    "pytest>=8.3.3",
    "ruff>=0.7.0",
]
